added more css and icons
