+++ src/components/01-intro-section/components/WhoAmI.jsx
@@ -1,5 +1,3 @@
-import { React } from "react"
-
 export const WhoAmI = () => {
   return (
     <>
