+++ src/components/01-intro-section/introSection.css
@@ -1,7 +1,3 @@
-.intro-section {
-  /* height: 100dvh; */
-}
-
 .hello {
   position: relative;
   /* background-color: red; */
