+++ src/components/02-tech-section/TechSection.jsx
