+++ src/components/03-featured-project-section/FeaturedProjectsSection.jsx
@@ -1,0 +1,18 @@
+import { ProjectCard } from "./components/ProjectCard"
+import data from "./data/card-data.json"
+import "./featured-project-section.css"
+
+
+export const FeaturedProjectSection = () => {
+  return (
+    <>
+      <section>
+        <h2>Featured Projects</h2>
+        {data.cardData.map((card, id) => (
+          <ProjectCard key={id} title={card.title} description={card.description} languageTag={card.languageTag} />
+        ))}
+        <button>See more projects</button>
+      </section>
+    </>
+  )
+}
