+++ src/components/03-featured-project-section/featured-project-section.css
@@ -1,0 +1,18 @@
+.language-tag-container {
+  display: flex;
+  justify-content: center;
+  gap: 5px;
+}
+
+.language-tag {
+  border: solid black 1px;
+  border-radius: 4px;
+  padding: 6px;
+}
+
+.web-btn {}
+
+.web-icon {
+  height: 28px;
+  width: auto;
+}
