+++ src/components/04-skills-section/components/SkillCategory.jsx
