+++ src/components/04-skills-section/data/skills-data.json
@@ -1,0 +1,32 @@
+{
+  "SkillsData": {
+    "Code": [
+      "HTML5",
+      "CSS3",
+      "JavaScript",
+      "ES6",
+      "React",
+      "Styled Components",
+      "GitHub"
+    ],
+    "Toolbox": [
+      "Atom",
+      "Postman",
+      "Adobe Photoshop",
+      "Adobe Illustrator",
+      "Figma",
+      "Keynote",
+      "Slack"
+    ],
+    "Upcoming": [
+      "Node.js"
+    ],
+    "More": [
+      "Branding",
+      "Strategy",
+      "Process Design",
+      "Concept Development",
+      "Agile methodology"
+    ]
+  }
+}
