+++ src/components/05-my-words/data/my-words-data.json
@@ -1,0 +1,10 @@
+{
+  "words": [
+    {
+      "id": 1,
+      "title": "From Freelance Filmmaker to Web Developer",
+      "text": "Hi everyone, I’ve been holding something in for a while now, and I feel like it’s time to share it with you all. I wanted to share the journey I’ve started to embark on and the reasons behind it. So here it goes!",
+      "img": "/simon2.webp"
+    }
+  ]
+}
