+++ src/components/05-my-words/my-words.css
