--- src/components/sections/03-featured-project-section/CodeUsed.jsx
@@ -1,11 +1,0 @@
-
-export const CodeUsed = () => {
-  return (
-    <>
-      <p>Lang 1</p>
-      <p>Lang 2</p>
-      <p>Lang 3</p>
-      <p>Lang 4</p>
-    </>
-  )
-}
