--- src/components/sections/04-skills-section/SkillsData.jsx
@@ -1,30 +1,0 @@
-export const SkillsData = {
-  Code: [
-    "HTML5",
-    "CSS3",
-    "JavaScript",
-    "ES6",
-    "React",
-    "Styled Components",
-    "GitHub"
-  ],
-  Toolbox: [
-    "Atom",
-    "Postman",
-    "Adobe Photoshop",
-    "Adobe Illustrator",
-    "Figma",
-    "Keynote",
-    "Slack"
-  ],
-  Upcoming: [
-    "Node.js"
-  ],
-  More: [
-    "Branding",
-    "Strategy",
-    "Process Design",
-    "Concept Development",
-    "Agile methodology"
-  ]
-};
