--- src/data/my-words-data.json
@@ -1,22 +1,0 @@
-{
-  "words": [
-    {
-      "id": 1,
-      "title": "From Freelance Filmmaker to Web Developer",
-      "text": "Hi everyone, I’ve been holding something in for a while now, and I feel like it’s time to share it with you all. I wanted to share the journey I’ve started to embark on and the reasons behind it. So here it goes!",
-      "img": "https://images.unsplash.com/photo-1557008075-7f2c5efa4cfd?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2497&q=80"
-    },
-    {
-      "id": 2,
-      "title": "From Freelance Filmmaker to Web Developer",
-      "text": "Hi everyone, I’ve been holding something in for a while now, and I feel like it’s time to share it with you all. I wanted to share the journey I’ve started to embark on and the reasons behind it. So here it goes!",
-      "img": "https://images.unsplash.com/photo-1557008075-7f2c5efa4cfd?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2497&q=80"
-    },
-    {
-      "id": 3,
-      "title": "From Freelance Filmmaker to Web Developer",
-      "text": "Hi everyone, I’ve been holding something in for a while now, and I feel like it’s time to share it with you all. I wanted to share the journey I’ve started to embark on and the reasons behind it. So here it goes!",
-      "img": "https://images.unsplash.com/photo-1557008075-7f2c5efa4cfd?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2497&q=80"
-    }
-  ]
-}
