+++ src/index.css
@@ -24,7 +24,7 @@ h1 {
 }
 
 h2 {
-  font-size: 24px;
+  font-size: 48px;
   font-weight: 500;
   font-style: normal;
 }
@@ -39,6 +39,22 @@ p {
   font-size: 20px;
   font-weight: 400;
   font-style: normal;
+}
+
+button {
+  width: 300px;
+  height: 48px;
+  border-radius: 12px;
+  margin: 5px;
+  border: none;
+  color: white;
+  background-color: black;
+  font-size: 18px;
+  display: inline-flex;
+  align-items: center;
+  gap: 12px;
+  padding: 8px 16px;
+  cursor: pointer;
 }
 
 section {
